added WS handler
